requirements.txt
numpy==1.24.3
torch==2.0.1
matplotlib==3.7.1
alpha_vantage==2.3.1
streamlit==1.22.0
pandas==1.5.3
requests==2.28.2
logging==0.5.1.2
python-dotenv==1.0.0